Create server() and server_graceful() functions, add get_bundle() test
